[REFACTOR] : Auth 하위 디렉토리 클래스 이동
