+++ module-api/src/main/kotlin/org/depromeet/team3/auth/controller/AuthController.kt
@@ -7,7 +7,7 @@ import io.swagger.v3.oas.annotations.tags.Tag
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
 import org.depromeet.team3.auth.application.AuthService
-import org.depromeet.team3.auth.application.response.UserProfileResponse
+import org.depromeet.team3.auth.dto.UserProfileResponse
 import org.depromeet.team3.common.ContextConstants
 import org.depromeet.team3.common.response.DpmApiResponse
 import org.springframework.web.bind.annotation.*
