+++ module-api/src/main/kotlin/org/depromeet/team3/security/SecurityConfig.kt
@@ -37,7 +37,8 @@ class SecurityConfig(
                 it.requestMatchers(
                     "/",
                     "/api/auth/**",
-                    "/auth/callback/**",  // 카카오 OAuth 콜백 경로 추가
+                    "/api/v1/auth/**",
+                    "/auth/callback/**",
                     "/swagger", "/swagger/", "/swagger-ui/**", "/v3/api-docs/**",
                     "/index.html", "/static/**", "/favicon.ico"
                 ).permitAll()
